CPU secitons IF, IFID, Control, ID, IDEX, EX hooked up
All of the CPU pipeline sections up through the Execute stage have been
updated and connected via the top level design Trosistor_CPU. Comments
are in each of the modules to figure out what's going on and how things
are connected together. Only sections we're missing are
Hazard_Detector/PC_control (which I'll connect up tomorrow) and then
Jake's sections after the Execute stage. I have a rough test bench in
the works for testing the top level, and other test benches for each
section are being written as well.